refactor: isolate FirebaseUI setup from LoginPresenter
