+++ frontend/src/presenters/loginPresenter.tsx
@@ -1,13 +1,11 @@
 import { useEffect } from "react";
-import { EmailAuthProvider, GoogleAuthProvider } from "firebase/auth";
-import "firebaseui/dist/firebaseui.css";
 import { LoginView } from "../views/loginView";
 import { useNavigate } from "react-router-dom";
 import { useAppDispatch, useAppSelector } from "../store/store";
 import { getAuthUserCB, getAuthLoadingCB, getCurrentLanguageCB } from "../store/selectors";
 import { Suspense } from "../components/Suspense";
-import { getLoginAuthUI, resetLoginAuthUI } from "../store/authThunks";
 import { translations } from "../utils/translations";
+import { startLoginAuthWidget } from "../firebase/loginAuthWidget";
 
 export function LoginPresenter() {
     const dispatch = useAppDispatch();
@@ -26,30 +24,12 @@ export function LoginPresenter() {
     useEffect(() => {
         if (loading || user) return;
 
-        const authUI = dispatch(getLoginAuthUI());
-
-        const uiConfig = {
-            signInFlow: "popup",
-            signInOptions: [GoogleAuthProvider.PROVIDER_ID, EmailAuthProvider.PROVIDER_ID],
-            callbacks: {
-                signInSuccessWithAuthResult: () => {
-                    navigate("/");
-                    return false; // Prevent FirebaseUI's default redirect
-                },
+        return startLoginAuthWidget({
+            dispatch,
+            onSignInSuccess: () => {
+                navigate("/");
             },
-        };
-
-        // Delay to ensure the container is rendered
-        const setupTimeout = setTimeout(() => {
-            if (document.getElementById("firebaseui-auth-container")) {
-                authUI.start("#firebaseui-auth-container", uiConfig);
-            }
-        }, 0);
-
-        return () => {
-            clearTimeout(setupTimeout);
-            dispatch(resetLoginAuthUI());
-        };
+        });
     }, [loading, user, navigate, dispatch]);
 
     if (loading) {
